#35: move parsing function into lib folder
